+++ parking_permits/tests/test_cron.py
@@ -1,7 +1,9 @@
 from datetime import datetime
 from datetime import timezone as dt_tz
+from decimal import Decimal
 from unittest.mock import patch
 
+from django.core import mail
 from django.test import TestCase
 from django.utils import timezone as tz
 from freezegun import freeze_time
@@ -11,15 +13,22 @@ from parking_permits.cron import (
     automatic_expiration_remind_notification_of_permits,
     automatic_remove_obsolete_customer_data,
     automatic_syncing_of_permits_to_parkkihubi,
+    handle_announcement_emails,
 )
-from parking_permits.models import Customer
+from parking_permits.models import Customer, Refund
+from parking_permits.models.order import OrderStatus
 from parking_permits.models.parking_permit import (
     ContractType,
     ParkingPermit,
+    ParkingPermitEndType,
     ParkingPermitStatus,
 )
+from parking_permits.tests.factories import ParkingZoneFactory
+from parking_permits.tests.factories.announcement import AnnouncementFactory
 from parking_permits.tests.factories.customer import CustomerFactory
+from parking_permits.tests.factories.order import OrderFactory, OrderItemFactory
 from parking_permits.tests.factories.parking_permit import ParkingPermitFactory
+from parking_permits.tests.factories.product import ProductFactory
 
 
 class CronTestCase(TestCase):
@@ -66,42 +75,100 @@ class CronTestCase(TestCase):
         automatic_syncing_of_permits_to_parkkihubi()
         mock_sync.assert_called()
 
-    @freeze_time(tz.make_aware(datetime(2023, 11, 30, 0, 22)))
+    @freeze_time(tz.make_aware(datetime(2024, 11, 30, 0, 22)))
     def test_automatic_expiration_permits(self):
         ParkingPermitFactory(
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 12, 1)),
+            start_time=tz.make_aware(datetime(2024, 11, 2, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 12, 1, 23, 59, 59)),
+            month_count=1,
+            contract_type=ContractType.FIXED_PERIOD,
             status=ParkingPermitStatus.VALID,
         )
         ParkingPermitFactory(
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 11, 29)),
+            start_time=tz.make_aware(datetime(2024, 10, 30, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 11, 29, 23, 59, 59)),
+            month_count=1,
+            contract_type=ContractType.FIXED_PERIOD,
             status=ParkingPermitStatus.DRAFT,
         )
         ParkingPermitFactory(
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 12, 1)),
+            start_time=tz.make_aware(datetime(2024, 11, 2, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 12, 1, 23, 59, 59)),
+            month_count=1,
+            contract_type=ContractType.FIXED_PERIOD,
             status=ParkingPermitStatus.DRAFT,
         )
-        ParkingPermitFactory(
+        closable_permit_1 = ParkingPermitFactory(
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 11, 29)),
+            start_time=tz.make_aware(datetime(2024, 10, 29, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 11, 28, 23, 59, 59)),
+            month_count=1,
+            contract_type=ContractType.FIXED_PERIOD,
+            status=ParkingPermitStatus.VALID,
+        )
+        closable_permit_2 = ParkingPermitFactory(
+            customer=self.customer,
+            start_time=tz.make_aware(datetime(2024, 9, 30, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 11, 29, 23, 59, 59)),
+            month_count=2,
+            contract_type=ContractType.FIXED_PERIOD,
+            status=ParkingPermitStatus.VALID,
+        )
+        closable_permit_3 = ParkingPermitFactory(
+            customer=self.customer,
+            start_time=tz.make_aware(datetime(2024, 10, 30, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 11, 29, 23, 59, 59)),
+            month_count=1,
+            contract_type=ContractType.OPEN_ENDED,
             status=ParkingPermitStatus.VALID,
         )
         ParkingPermitFactory(
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 11, 29)),
+            start_time=tz.make_aware(datetime(2024, 10, 30, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 11, 29, 23, 59, 59)),
+            month_count=1,
+            contract_type=ContractType.FIXED_PERIOD,
             status=ParkingPermitStatus.CLOSED,
         )
         valid_permits = ParkingPermit.objects.filter(status=ParkingPermitStatus.VALID)
         draft_permits = ParkingPermit.objects.filter(status=ParkingPermitStatus.DRAFT)
-        self.assertEqual(valid_permits.count(), 2)
-        self.assertEqual(draft_permits.count(), 2)
-        automatic_expiration_of_permits()
         closed_permits = ParkingPermit.objects.filter(status=ParkingPermitStatus.CLOSED)
+        self.assertEqual(valid_permits.count(), 4)
+        self.assertEqual(draft_permits.count(), 2)
+        self.assertEqual(closed_permits.count(), 1)
+        automatic_expiration_of_permits()
         self.assertEqual(valid_permits.count(), 1)
         self.assertEqual(draft_permits.count(), 0)
-        self.assertEqual(closed_permits.count(), 2)
+        self.assertEqual(closed_permits.count(), 4)
+        closable_permit_1.refresh_from_db()
+        self.assertEqual(
+            closable_permit_1.end_time,
+            tz.make_aware(datetime(2024, 11, 29, 23, 59, 59, 999999)),
+        )
+        self.assertEqual(
+            closable_permit_1.end_type, ParkingPermitEndType.PREVIOUS_DAY_END
+        )
+        closable_permit_2.refresh_from_db()
+        self.assertEqual(
+            closable_permit_2.end_time,
+            tz.make_aware(datetime(2024, 11, 29, 23, 59, 59, 999999)),
+        )
+        self.assertEqual(
+            closable_permit_2.end_type, ParkingPermitEndType.PREVIOUS_DAY_END
+        )
+        closable_permit_3.refresh_from_db()
+        self.assertEqual(
+            closable_permit_3.end_time,
+            tz.make_aware(datetime(2024, 11, 29, 23, 59, 59, 999999)),
+        )
+        self.assertEqual(
+            closable_permit_3.end_type, ParkingPermitEndType.PREVIOUS_DAY_END
+        )
+        # Make sure that no refunds were created for these cases
+        self.assertEqual(Refund.objects.count(), 0)
 
     @freeze_time(tz.make_aware(datetime(2023, 11, 30, 0, 22)))
     def test_automatic_expiration_permits_with_primary_permit_change(self):
@@ -148,32 +215,57 @@ class CronTestCase(TestCase):
             tz.make_aware(datetime(2023, 11, 29, 23, 59, 59, 999999)),
         )
 
-    @freeze_time(tz.make_aware(datetime(2023, 11, 30, 0, 22)))
+    @freeze_time(tz.make_aware(datetime(2024, 11, 30, 0, 22)))
     def test_automatic_expiration_permits_with_permit_vehicle_changed(self):
-        ParkingPermitFactory(
+        permit_1 = ParkingPermitFactory(
             id=80000010,
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 12, 29, 23, 59)),
+            start_time=tz.make_aware(datetime(2024, 10, 30, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 12, 29, 23, 59, 59)),
             vehicle_changed=True,
-            vehicle_changed_date=tz.make_aware(datetime(2023, 11, 29)).date(),
+            vehicle_changed_date=tz.make_aware(datetime(2024, 11, 29)).date(),
+            month_count=2,
+            contract_type=ContractType.FIXED_PERIOD,
             status=ParkingPermitStatus.VALID,
             primary_vehicle=True,
         )
+        unit_price = Decimal(67.5)
+        product = ProductFactory(
+            unit_price=unit_price,
+        )
+        order = OrderFactory(
+            customer=self.customer, status=OrderStatus.CONFIRMED, talpa_order_id=None
+        )
+        OrderItemFactory(
+            order=order,
+            permit=permit_1,
+            start_time=tz.make_aware(datetime(2024, 10, 30, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 12, 29, 23, 59, 59)),
+            product=product,
+            unit_price=unit_price,
+            payment_unit_price=unit_price,
+            quantity=2,
+            vat=Decimal(0.255),
+        )
+        permit_1.orders.add(order)
         ParkingPermitFactory(
             id=80000020,
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 12, 29)),
+            start_time=tz.make_aware(datetime(2024, 10, 30, 10, 0, 0)),
+            end_time=tz.make_aware(datetime(2024, 12, 29, 23, 59, 59)),
+            month_count=2,
+            contract_type=ContractType.FIXED_PERIOD,
             status=ParkingPermitStatus.VALID,
             primary_vehicle=False,
         )
         ParkingPermitFactory(
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 11, 29)),
+            end_time=tz.make_aware(datetime(2024, 11, 29)),
             status=ParkingPermitStatus.DRAFT,
         )
         ParkingPermitFactory(
             customer=self.customer,
-            end_time=tz.make_aware(datetime(2023, 12, 1)),
+            end_time=tz.make_aware(datetime(2024, 12, 1)),
             status=ParkingPermitStatus.DRAFT,
         )
         valid_permits = ParkingPermit.objects.filter(status=ParkingPermitStatus.VALID)
@@ -191,12 +283,23 @@ class CronTestCase(TestCase):
         self.assertEqual(draft_permits.all().count(), 0)
         closed_permits = closed_permits.all()
         self.assertEqual(closed_permits.count(), 1)
-        closed_permit = closed_permits.first()
+        permit_1.refresh_from_db()
+        closed_permit = permit_1
         self.assertEqual(closed_permit.id, 80000010)
+        self.assertEqual(closed_permit.vehicle_changed, True)
+        self.assertEqual(
+            closed_permit.vehicle_changed_date,
+            tz.make_aware(datetime(2024, 11, 29)).date(),
+        )
+        self.assertEqual(closed_permit.primary_vehicle, True)
+        self.assertEqual(closed_permit.end_type, ParkingPermitEndType.PREVIOUS_DAY_END)
         self.assertEqual(
             closed_permit.end_time,
-            tz.make_aware(datetime(2023, 11, 29, 23, 59, 59, 999999)),
+            tz.make_aware(datetime(2024, 11, 29, 23, 59, 59, 999999)),
         )
+        self.assertEqual(Refund.objects.count(), 1)
+        refund = Refund.objects.first()
+        self.assertEqual(refund.amount, Decimal(67.5))  # Should be one month's amount
 
 
 class AutomaticRemoveObsoleteCustomerDataTestCase(TestCase):
@@ -271,3 +374,83 @@ class AutomaticExpirationRemindPermitNotificationTestCase(TestCase):
         self.assertEqual(valid_permits.count(), 5)
         expiring_permits = automatic_expiration_remind_notification_of_permits()
         self.assertEqual(expiring_permits.count(), 2)
+
+
+class AnnouncementEmailHandlingTestCase(TestCase):
+    def setUp(self):
+        self.zone_a = ParkingZoneFactory(name="A")
+        self.zone_b = ParkingZoneFactory(name="B")
+        self.zone_c = ParkingZoneFactory(name="C")
+        self.announcement = AnnouncementFactory(subject_en="Test announcement")
+        self.customer_1 = CustomerFactory(email="test@mail.com")
+        self.customer_2 = CustomerFactory(email="test2@mail.com")
+        self.customer_3 = CustomerFactory(email="test3@mail.com")
+
+        self.customer_1_permit_1 = ParkingPermitFactory(
+            customer=self.customer_1,
+            parking_zone=self.zone_a,
+            status=ParkingPermitStatus.VALID,
+        )
+        self.customer_1_permit_2 = ParkingPermitFactory(
+            customer=self.customer_1,
+            parking_zone=self.zone_a,
+            status=ParkingPermitStatus.DRAFT,
+        )
+        self.customer_1_permit_3 = ParkingPermitFactory(
+            customer=self.customer_1,
+            parking_zone=self.zone_c,
+            status=ParkingPermitStatus.CANCELLED,
+        )
+        self.customer_1_permit_4 = ParkingPermitFactory(
+            customer=self.customer_1,
+            parking_zone=self.zone_a,
+            status=ParkingPermitStatus.VALID,
+        )
+
+        self.customer_2_permit_1 = ParkingPermitFactory(
+            customer=self.customer_2,
+            parking_zone=self.zone_b,
+            status=ParkingPermitStatus.DRAFT,
+        )
+        self.customer_2_permit_2 = ParkingPermitFactory(
+            customer=self.customer_2,
+            parking_zone=self.zone_b,
+            status=ParkingPermitStatus.VALID,
+        )
+
+        self.customer_3_permit_1 = ParkingPermitFactory(
+            customer=self.customer_3,
+            parking_zone=self.zone_c,
+            status=ParkingPermitStatus.DRAFT,
+        )
+
+    def test_handle_announcement_emails(self):
+        self.announcement._parking_zones.set([self.zone_a])
+        self.assertEqual(self.announcement.emails_handled, False)
+        handle_announcement_emails()
+        self.assertEqual(len(mail.outbox), 1)
+        self.assertTrue("Test announcement" in mail.outbox[0].subject)
+        self.announcement.refresh_from_db()
+        self.assertEqual(self.announcement.emails_handled, True)
+
+    def test_multizone_announcement(self):
+        self.announcement._parking_zones.set([self.zone_a, self.zone_b])
+        handle_announcement_emails()
+        self.assertEqual(len(mail.outbox), 2)
+        self.assertTrue("Test announcement" in mail.outbox[0].subject)
+        self.announcement.refresh_from_db()
+        self.assertEqual(self.announcement.emails_handled, True)
+
+    def test_valid_permits_only(self):
+        self.announcement._parking_zones.set([self.zone_a, self.zone_b, self.zone_c])
+        handle_announcement_emails()
+        self.assertEqual(len(mail.outbox), 2)
+        self.announcement.refresh_from_db()
+        self.assertEqual(self.announcement.emails_handled, True)
+
+    def test_no_valid_permits(self):
+        self.announcement._parking_zones.set([self.zone_c])
+        handle_announcement_emails()
+        self.assertEqual(len(mail.outbox), 0)
+        self.announcement.refresh_from_db()
+        self.assertEqual(self.announcement.emails_handled, True)